requirements.txt
matplotlib
numpy
Pillow
tensorboardX
torch
torchvision
